Cross-compile most of tests in testsuite.compiler.RegressionTest.
`should_support_class_literals_for_existential_value_types_issue_218`
should pass, which is why it is in the shared part, but it doesn't
due to not being bug-compatible with the JVM. This is filed as
